requirements.txt
Pillow~=9.5.0
pip~=23.1
wheel~=0.40.0
pytz~=2023.3
setuptools~=67.6.1
Django~=4.2
asgiref~=3.6.0
sqlparse~=0.4.3
requests~=2.28.2
certifi~=2022.12.7
Markdown~=3.4.3
idna~=3.4
urllib3~=1.26.15
zipp~=3.15.0
django-filter~=23.1
djangorestframework~=3.14.0
PyJWT~=2.6.0
djangorestframework-simplejwt~=5.2.2
mysqlclient~=2.1.1